fix(syscall/proc): i32 in sys_exit and sys_spawn
fix(userspace/string.h): gets_s now correctly work with \b

chore(kmain): now ring3 is default when pressing any key (instead of ksh)

feat(termosh): spawn in termosh

+++ kernel/inc/syscalls/proc.h
@@ -5,6 +5,6 @@
 
 #include <types.h>
 
-u64 sys_exit(i32 code);
-u64 sys_spawn(const char* path);
-u64 sys_wait(u64 pid);
+i32 sys_exit(i32 code);
+i32 sys_spawn(const char* path);
+i32 sys_wait(u64 pid);
